chore(cli): split doctor.py into checks + plugin_repair modules (#1270)
doctor.py was 667 LOC with ~20 module-level functions mixing three
unrelated concerns. Split internal helpers into focused private modules
without changing the public CLI surface:

- src/teatree/cli/doctor.py (522 LOC) — DoctorService, IntrospectionHelpers,
  pyproject helpers used by make_editable/restore_sources, and the check()
  Typer entry point.
- src/teatree/cli/_doctor_checks.py (99 LOC) — _check_single_db,
  _check_singletons, _check_editable_sanity, _check_skills,
  _check_legacy_overlay_alias.
- src/teatree/cli/_doctor_plugin_repair.py (131 LOC) — _read_json_safe,
  _resolve_main_clone, _repair_marketplace_json, _repair_installed_plugins,
  _repair_enabled_plugins, _ensure_plugin_registered,
  _do_ensure_plugin_registered.

Each new file is under 300 LOC and under 10 module-level functions.

Import-migration choice: re-export every moved helper from teatree.cli.doctor
via `from ._doctor_checks import ...` / `from ._doctor_plugin_repair import ...`
and an explicit __all__. Existing consumers (tests/cli_doctor/*.py reaching
for teatree_cli_doctor._find_host_project_root, ._write_dev_sources_marker,
._resolve_main_clone, ._check_legacy_overlay_alias; src/teatree/cli/setup.py;
src/teatree/cli/info.py; src/teatree/cli/agent.py; tests using
patch("teatree.cli.doctor._find_host_project_root", ...)) keep working
unchanged — no consumer-side import update needed.

Two cross-module references avoided through deferred imports:
_doctor_checks._check_editable_sanity calls DoctorService.check_editable_sanity()
and _doctor_plugin_repair._resolve_main_clone calls
DoctorService.find_teatree_repo(). Both import DoctorService at call time
to avoid circular imports at module load.

Closes #1270
